[570] Moving the helper functions to the `mqtt_tests` crate
