+++ crates/tests/mqtt_tests/src/lib.rs
@@ -1,6 +1,9 @@
+mod message_streams;
 mod test_mqtt_client;
 pub mod test_mqtt_server;
 pub mod with_timeout;
 
-pub use test_mqtt_client::{assert_received, publish};
+pub use message_streams::*;
+pub use test_mqtt_client::assert_received;
+pub use test_mqtt_client::publish;
 pub use test_mqtt_server::test_mqtt_broker;
